Extract repr and matches utils to separate modules
